requirements.txt
opensearch-py
sentence-transformers
torch
pandas
faiss-cpu
rank_bm25
requests
google-generativeai